backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
numpy==2.1.3
soundfile==0.12.1
faster-whisper==1.0.0
pandas
joblib
scikit-learn
librosa
python-multipart
websockets
python-dotenv
sqlalchemy
pymysql